Final review response I hope)
